Improved Safari support
For whatever reason, it is possible for Safari to get scared of its own `::-webkit-...` rule declarations.

We don't really need any of those, as those pseudo elements aren't going to end up in the final source anyway.

So this patch treats any error when testing selector matches as a non-match.

Fixes #1